[TS-64] Rename: Chat -> Chatting
