+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 
 import ChartPage from "@/pages/ChartPage";
-import ChatPage from "@/pages/ChatPage";
+import ChattingPage from "@/pages/ChattingPage";
 import HabitDeletePage from "@/pages/HabitDeletePage";
 import HomePage from "@/pages/HomePage";
 import MyHabitPage from "@/pages/MyHabitPage";
@@ -32,7 +32,7 @@ const App = () => {
 				<Route path="/notification-setting" element={<NotificationSettingPage />} />
 				<Route path="/withdrawal" element={<WithdrawalPage />} />
 				<Route path="/rest-record" element={<RestRecordPage />} />
-				<Route path="/chat" element={<ChatPage />} />
+				<Route path="/chatting" element={<ChattingPage />} />
 				<Route path="/habit-delete" element={<HabitDeletePage />} />
 			</Routes>
 		</BrowserRouter>
