--- src/pages/ChatPage.tsx
@@ -1,11 +1,0 @@
-import LocationModal from "@/components/OnBoarding/LocationModal";
-
-function ChatPage() {
-	return (
-		<div>
-			<LocationModal />
-		</div>
-	);
-}
-
-export default ChatPage;
