+++ src/pages/ChattingPage.tsx
@@ -1,0 +1,11 @@
+import LocationModal from "@/components/Chatting/LocationModal";
+
+function ChattingPage() {
+	return (
+		<div>
+			<LocationModal />
+		</div>
+	);
+}
+
+export default ChattingPage;
